+++ Energy.Solving/PathFinding/Node.cs
@@ -5,7 +5,7 @@ namespace Energy.Solving.PathFinding
         public int? Component { get; set; }
 
         public double? Distance { get; set; }
-        public int? Conduction { get; set; }
+        public double? Conduction { get; set; }
 
         public string Name { get; }
